# template/api/system/v1/role_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.31.1
// source: system/v1/role.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationRoleCheckAPIPermission = "/api.system.v1.Role/CheckAPIPermission"
const OperationRoleCreateRole = "/api.system.v1.Role/CreateRole"
const OperationRoleDeleteRole = "/api.system.v1.Role/DeleteRole"
const OperationRoleGetRole = "/api.system.v1.Role/GetRole"
const OperationRoleGetUserPermission = "/api.system.v1.Role/GetUserPermission"
const OperationRoleListRole = "/api.system.v1.Role/ListRole"
const OperationRoleListRoleMenus = "/api.system.v1.Role/ListRoleMenus"
const OperationRoleListRolePermission = "/api.system.v1.Role/ListRolePermission"
const OperationRoleListUserRoles = "/api.system.v1.Role/ListUserRoles"
const OperationRoleUpdateRole = "/api.system.v1.Role/UpdateRole"
const OperationRoleUpdateRoleDataScope = "/api.system.v1.Role/UpdateRoleDataScope"
const OperationRoleUpdateRoleMenus = "/api.system.v1.Role/UpdateRoleMenus"
const OperationRoleUpdateUserRole = "/api.system.v1.Role/UpdateUserRole"

type RoleHTTPServer interface {
	// CheckAPIPermission CheckAPIPermission checks if a role has permission to access a specific API
	// endpoint
	CheckAPIPermission(context.Context, *CheckAPIPermissionRequest) (*CheckAPIPermissionReply, error)
	// CreateRole CreateRole creates a new role with specified attributes and menu
	// permissions
	CreateRole(context.Context, *CreateRoleRequest) (*CreateRoleReply, error)
	// DeleteRole DeleteRole deletes a role by its ID
	DeleteRole(context.Context, *DeleteRoleRequest) (*DeleteRoleReply, error)
	// GetRole GetRole retrieves detailed information about a specific role
	GetRole(context.Context, *GetRoleRequest) (*GetRoleReply, error)
	// GetUserPermission GetUserPermission retrieves all permissions for a specific user
	GetUserPermission(context.Context, *GetUserPermissionRequest) (*GetUserPermissionReply, error)
	// ListRole ListRole retrieves a paginated list of roles with optional filtering
	ListRole(context.Context, *ListRoleRequest) (*ListRoleReply, error)
	// ListRoleMenus ListRoleMenus retrieves the list of menu IDs assigned to a specific role
	ListRoleMenus(context.Context, *ListRoleMenusRequest) (*ListRoleMenusReply, error)
	// ListRolePermission ListRolePermission retrieves all menu permissions assigned to a specific
	// role
	ListRolePermission(context.Context, *ListRolePermissionRequest) (*ListRolePermissionReply, error)
	// ListUserRoles ListUserRoles retrieves the list of role IDs assigned to a specific user
	ListUserRoles(context.Context, *ListUserRolesRequest) (*ListUserRolesReply, error)
	// UpdateRole UpdateRole updates an existing role's attributes and menu permissions
	UpdateRole(context.Context, *UpdateRoleRequest) (*UpdateRoleReply, error)
	// UpdateRoleDataScope UpdateRoleDataScope updates the data scope permissions for a specific role
	UpdateRoleDataScope(context.Context, *UpdateRoleDataScopeRequest) (*UpdateRoleDataScopeReply, error)
	// UpdateRoleMenus UpdateRoleMenus updates the menu permissions for a specific role
	UpdateRoleMenus(context.Context, *UpdateRoleMenusRequest) (*UpdateRoleMenusReply, error)
	// UpdateUserRole UpdateUserRole updates the roles assigned to a specific user
	UpdateUserRole(context.Context, *UpdateUserRoleRequest) (*UpdateUserRoleReply, error)
}

func RegisterRoleHTTPServer(s *http.Server, srv RoleHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/system/role", _Role_CreateRole0_HTTP_Handler(srv))
	r.POST("/api/v1/system/role/update", _Role_UpdateRole0_HTTP_Handler(srv))
	r.POST("/api/v1/system/role/delete", _Role_DeleteRole0_HTTP_Handler(srv))
	r.GET("/api/v1/system/role/get", _Role_GetRole0_HTTP_Handler(srv))
	r.GET("/api/v1/system/role/list", _Role_ListRole0_HTTP_Handler(srv))
	r.GET("/api/v1/system/role/perm", _Role_ListRolePermission0_HTTP_Handler(srv))
	r.POST("/api/v1/system/role/menu", _Role_UpdateRoleMenus0_HTTP_Handler(srv))
	r.POST("/api/v1/system/role/scope", _Role_UpdateRoleDataScope0_HTTP_Handler(srv))
	r.POST("/api/v1/system/role/user", _Role_UpdateUserRole0_HTTP_Handler(srv))
	r.GET("/api/v1/system/role/menus", _Role_ListRoleMenus0_HTTP_Handler(srv))
	r.GET("/api/v1/system/role/users", _Role_ListUserRoles0_HTTP_Handler(srv))
	r.POST("/api/v1/system/role/check", _Role_CheckAPIPermission0_HTTP_Handler(srv))
	r.GET("/api/v1/system/role/user/permission", _Role_GetUserPermission0_HTTP_Handler(srv))
}

func _Role_CreateRole0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleCreateRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateRole(ctx, req.(*CreateRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateRoleReply)
		return ctx.Result(200, reply)
	}
}

func _Role_UpdateRole0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleUpdateRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateRole(ctx, req.(*UpdateRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateRoleReply)
		return ctx.Result(200, reply)
	}
}

func _Role_DeleteRole0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleDeleteRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteRole(ctx, req.(*DeleteRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteRoleReply)
		return ctx.Result(200, reply)
	}
}

func _Role_GetRole0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetRoleRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleGetRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetRole(ctx, req.(*GetRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetRoleReply)
		return ctx.Result(200, reply)
	}
}

func _Role_ListRole0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRoleRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleListRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListRole(ctx, req.(*ListRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListRoleReply)
		return ctx.Result(200, reply)
	}
}

func _Role_ListRolePermission0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRolePermissionRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleListRolePermission)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListRolePermission(ctx, req.(*ListRolePermissionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListRolePermissionReply)
		return ctx.Result(200, reply)
	}
}

func _Role_UpdateRoleMenus0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateRoleMenusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleUpdateRoleMenus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateRoleMenus(ctx, req.(*UpdateRoleMenusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateRoleMenusReply)
		return ctx.Result(200, reply)
	}
}

func _Role_UpdateRoleDataScope0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateRoleDataScopeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleUpdateRoleDataScope)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateRoleDataScope(ctx, req.(*UpdateRoleDataScopeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateRoleDataScopeReply)
		return ctx.Result(200, reply)
	}
}

func _Role_UpdateUserRole0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateUserRoleRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleUpdateUserRole)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateUserRole(ctx, req.(*UpdateUserRoleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateUserRoleReply)
		return ctx.Result(200, reply)
	}
}

func _Role_ListRoleMenus0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRoleMenusRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleListRoleMenus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListRoleMenus(ctx, req.(*ListRoleMenusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListRoleMenusReply)
		return ctx.Result(200, reply)
	}
}

func _Role_ListUserRoles0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListUserRolesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleListUserRoles)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListUserRoles(ctx, req.(*ListUserRolesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListUserRolesReply)
		return ctx.Result(200, reply)
	}
}

func _Role_CheckAPIPermission0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CheckAPIPermissionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleCheckAPIPermission)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CheckAPIPermission(ctx, req.(*CheckAPIPermissionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CheckAPIPermissionReply)
		return ctx.Result(200, reply)
	}
}

func _Role_GetUserPermission0_HTTP_Handler(srv RoleHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUserPermissionRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRoleGetUserPermission)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserPermission(ctx, req.(*GetUserPermissionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUserPermissionReply)
		return ctx.Result(200, reply)
	}
}

type RoleHTTPClient interface {
	// CheckAPIPermission CheckAPIPermission checks if a role has permission to access a specific API
	// endpoint
	CheckAPIPermission(ctx context.Context, req *CheckAPIPermissionRequest, opts ...http.CallOption) (rsp *CheckAPIPermissionReply, err error)
	// CreateRole CreateRole creates a new role with specified attributes and menu
	// permissions
	CreateRole(ctx context.Context, req *CreateRoleRequest, opts ...http.CallOption) (rsp *CreateRoleReply, err error)
	// DeleteRole DeleteRole deletes a role by its ID
	DeleteRole(ctx context.Context, req *DeleteRoleRequest, opts ...http.CallOption) (rsp *DeleteRoleReply, err error)
	// GetRole GetRole retrieves detailed information about a specific role
	GetRole(ctx context.Context, req *GetRoleRequest, opts ...http.CallOption) (rsp *GetRoleReply, err error)
	// GetUserPermission GetUserPermission retrieves all permissions for a specific user
	GetUserPermission(ctx context.Context, req *GetUserPermissionRequest, opts ...http.CallOption) (rsp *GetUserPermissionReply, err error)
	// ListRole ListRole retrieves a paginated list of roles with optional filtering
	ListRole(ctx context.Context, req *ListRoleRequest, opts ...http.CallOption) (rsp *ListRoleReply, err error)
	// ListRoleMenus ListRoleMenus retrieves the list of menu IDs assigned to a specific role
	ListRoleMenus(ctx context.Context, req *ListRoleMenusRequest, opts ...http.CallOption) (rsp *ListRoleMenusReply, err error)
	// ListRolePermission ListRolePermission retrieves all menu permissions assigned to a specific
	// role
	ListRolePermission(ctx context.Context, req *ListRolePermissionRequest, opts ...http.CallOption) (rsp *ListRolePermissionReply, err error)
	// ListUserRoles ListUserRoles retrieves the list of role IDs assigned to a specific user
	ListUserRoles(ctx context.Context, req *ListUserRolesRequest, opts ...http.CallOption) (rsp *ListUserRolesReply, err error)
	// UpdateRole UpdateRole updates an existing role's attributes and menu permissions
	UpdateRole(ctx context.Context, req *UpdateRoleRequest, opts ...http.CallOption) (rsp *UpdateRoleReply, err error)
	// UpdateRoleDataScope UpdateRoleDataScope updates the data scope permissions for a specific role
	UpdateRoleDataScope(ctx context.Context, req *UpdateRoleDataScopeRequest, opts ...http.CallOption) (rsp *UpdateRoleDataScopeReply, err error)
	// UpdateRoleMenus UpdateRoleMenus updates the menu permissions for a specific role
	UpdateRoleMenus(ctx context.Context, req *UpdateRoleMenusRequest, opts ...http.CallOption) (rsp *UpdateRoleMenusReply, err error)
	// UpdateUserRole UpdateUserRole updates the roles assigned to a specific user
	UpdateUserRole(ctx context.Context, req *UpdateUserRoleRequest, opts ...http.CallOption) (rsp *UpdateUserRoleReply, err error)
}

type RoleHTTPClientImpl struct {
	cc *http.Client
}

func NewRoleHTTPClient(client *http.Client) RoleHTTPClient {
	return &RoleHTTPClientImpl{client}
}

// CheckAPIPermission CheckAPIPermission checks if a role has permission to access a specific API
// endpoint
func (c *RoleHTTPClientImpl) CheckAPIPermission(ctx context.Context, in *CheckAPIPermissionRequest, opts ...http.CallOption) (*CheckAPIPermissionReply, error) {
	var out CheckAPIPermissionReply
	pattern := "/api/v1/system/role/check"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleCheckAPIPermission))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateRole CreateRole creates a new role with specified attributes and menu
// permissions
func (c *RoleHTTPClientImpl) CreateRole(ctx context.Context, in *CreateRoleRequest, opts ...http.CallOption) (*CreateRoleReply, error) {
	var out CreateRoleReply
	pattern := "/api/v1/system/role"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleCreateRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteRole DeleteRole deletes a role by its ID
func (c *RoleHTTPClientImpl) DeleteRole(ctx context.Context, in *DeleteRoleRequest, opts ...http.CallOption) (*DeleteRoleReply, error) {
	var out DeleteRoleReply
	pattern := "/api/v1/system/role/delete"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleDeleteRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetRole GetRole retrieves detailed information about a specific role
func (c *RoleHTTPClientImpl) GetRole(ctx context.Context, in *GetRoleRequest, opts ...http.CallOption) (*GetRoleReply, error) {
	var out GetRoleReply
	pattern := "/api/v1/system/role/get"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleGetRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetUserPermission GetUserPermission retrieves all permissions for a specific user
func (c *RoleHTTPClientImpl) GetUserPermission(ctx context.Context, in *GetUserPermissionRequest, opts ...http.CallOption) (*GetUserPermissionReply, error) {
	var out GetUserPermissionReply
	pattern := "/api/v1/system/role/user/permission"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleGetUserPermission))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListRole ListRole retrieves a paginated list of roles with optional filtering
func (c *RoleHTTPClientImpl) ListRole(ctx context.Context, in *ListRoleRequest, opts ...http.CallOption) (*ListRoleReply, error) {
	var out ListRoleReply
	pattern := "/api/v1/system/role/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleListRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListRoleMenus ListRoleMenus retrieves the list of menu IDs assigned to a specific role
func (c *RoleHTTPClientImpl) ListRoleMenus(ctx context.Context, in *ListRoleMenusRequest, opts ...http.CallOption) (*ListRoleMenusReply, error) {
	var out ListRoleMenusReply
	pattern := "/api/v1/system/role/menus"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleListRoleMenus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListRolePermission ListRolePermission retrieves all menu permissions assigned to a specific
// role
func (c *RoleHTTPClientImpl) ListRolePermission(ctx context.Context, in *ListRolePermissionRequest, opts ...http.CallOption) (*ListRolePermissionReply, error) {
	var out ListRolePermissionReply
	pattern := "/api/v1/system/role/perm"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleListRolePermission))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListUserRoles ListUserRoles retrieves the list of role IDs assigned to a specific user
func (c *RoleHTTPClientImpl) ListUserRoles(ctx context.Context, in *ListUserRolesRequest, opts ...http.CallOption) (*ListUserRolesReply, error) {
	var out ListUserRolesReply
	pattern := "/api/v1/system/role/users"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRoleListUserRoles))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateRole UpdateRole updates an existing role's attributes and menu permissions
func (c *RoleHTTPClientImpl) UpdateRole(ctx context.Context, in *UpdateRoleRequest, opts ...http.CallOption) (*UpdateRoleReply, error) {
	var out UpdateRoleReply
	pattern := "/api/v1/system/role/update"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleUpdateRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateRoleDataScope UpdateRoleDataScope updates the data scope permissions for a specific role
func (c *RoleHTTPClientImpl) UpdateRoleDataScope(ctx context.Context, in *UpdateRoleDataScopeRequest, opts ...http.CallOption) (*UpdateRoleDataScopeReply, error) {
	var out UpdateRoleDataScopeReply
	pattern := "/api/v1/system/role/scope"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleUpdateRoleDataScope))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateRoleMenus UpdateRoleMenus updates the menu permissions for a specific role
func (c *RoleHTTPClientImpl) UpdateRoleMenus(ctx context.Context, in *UpdateRoleMenusRequest, opts ...http.CallOption) (*UpdateRoleMenusReply, error) {
	var out UpdateRoleMenusReply
	pattern := "/api/v1/system/role/menu"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleUpdateRoleMenus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateUserRole UpdateUserRole updates the roles assigned to a specific user
func (c *RoleHTTPClientImpl) UpdateUserRole(ctx context.Context, in *UpdateUserRoleRequest, opts ...http.CallOption) (*UpdateUserRoleReply, error) {
	var out UpdateUserRoleReply
	pattern := "/api/v1/system/role/user"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRoleUpdateUserRole))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}