# template/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/system/role:
        post:
            tags:
                - Role
            description: |-
                CreateRole creates a new role with specified attributes and menu
                 permissions
            operationId: Role_CreateRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.CreateRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.CreateRoleReply'
    /api/v1/system/role/check:
        post:
            tags:
                - Role
            description: |-
                CheckAPIPermission checks if a role has permission to access a specific API
                 endpoint
            operationId: Role_CheckAPIPermission
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.CheckAPIPermissionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.CheckAPIPermissionReply'
    /api/v1/system/role/delete:
        post:
            tags:
                - Role
            description: DeleteRole deletes a role by its ID
            operationId: Role_DeleteRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.DeleteRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.DeleteRoleReply'
    /api/v1/system/role/get:
        get:
            tags:
                - Role
            description: GetRole retrieves detailed information about a specific role
            operationId: Role_GetRole
            parameters:
                - name: id
                  in: query
                  description: 角色ID
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.GetRoleReply'
    /api/v1/system/role/list:
        get:
            tags:
                - Role
            description: ListRole retrieves a paginated list of roles with optional filtering
            operationId: Role_ListRole
            parameters:
                - name: pageSize
                  in: query
                  description: 每页数量
                  schema:
                    type: integer
                    format: int32
                - name: current
                  in: query
                  description: 页码
                  schema:
                    type: integer
                    format: int32
                - name: name
                  in: query
                  description: 角色名称，用于筛选
                  schema:
                    type: string
                - name: status
                  in: query
                  description: 状态筛选 0 停用 1 正常
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListRoleReply'
    /api/v1/system/role/menu:
        post:
            tags:
                - Role
            description: UpdateRoleMenus updates the menu permissions for a specific role
            operationId: Role_UpdateRoleMenus
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.UpdateRoleMenusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.UpdateRoleMenusReply'
    /api/v1/system/role/menus:
        get:
            tags:
                - Role
            description: ListRoleMenus retrieves the list of menu IDs assigned to a specific role
            operationId: Role_ListRoleMenus
            parameters:
                - name: roleId
                  in: query
                  description: 角色ID
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListRoleMenusReply'
    /api/v1/system/role/perm:
        get:
            tags:
                - Role
            description: |-
                ListRolePermission retrieves all menu permissions assigned to a specific
                 role
            operationId: Role_ListRolePermission
            parameters:
                - name: roleId
                  in: query
                  description: 角色ID
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListRolePermissionReply'
    /api/v1/system/role/scope:
        post:
            tags:
                - Role
            description: UpdateRoleDataScope updates the data scope permissions for a specific role
            operationId: Role_UpdateRoleDataScope
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.UpdateRoleDataScopeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.UpdateRoleDataScopeReply'
    /api/v1/system/role/update:
        post:
            tags:
                - Role
            description: UpdateRole updates an existing role's attributes and menu permissions
            operationId: Role_UpdateRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.UpdateRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.UpdateRoleReply'
    /api/v1/system/role/user:
        post:
            tags:
                - Role
            description: UpdateUserRole updates the roles assigned to a specific user
            operationId: Role_UpdateUserRole
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.UpdateUserRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.UpdateUserRoleReply'
    /api/v1/system/role/user/permission:
        get:
            tags:
                - Role
            description: GetUserPermission retrieves all permissions for a specific user
            operationId: Role_GetUserPermission
            parameters:
                - name: userId
                  in: query
                  description: 用户ID
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.GetUserPermissionReply'
    /api/v1/system/role/users:
        get:
            tags:
                - Role
            description: ListUserRoles retrieves the list of role IDs assigned to a specific user
            operationId: Role_ListUserRoles
            parameters:
                - name: userId
                  in: query
                  description: 用户ID
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListUserRolesReply'
...
components:
    schemas:
...
        api.system.v1.CheckAPIPermissionReply:
            type: object
            properties:
                hasPermission:
                    type: boolean
                    description: 是否有权限
        api.system.v1.CheckAPIPermissionRequest:
            type: object
            properties:
                roleId:
                    type: string
                    description: 角色ID
                path:
                    type: string
                    description: API路径
                method:
                    type: string
                    description: HTTP方法
...
        api.system.v1.CreateRoleReply:
            type: object
            properties: {}
        api.system.v1.CreateRoleRequest:
            type: object
            properties:
                name:
                    type: string
                    description: 角色名称
                sort:
                    type: integer
                    description: 显示顺序
                    format: int32
                status:
                    type: integer
                    description: 状态(0停用 1正常)
                    format: int32
                dataScope:
                    type: integer
                    description: |-
                        数据范围(1:全部数据权限 2:自定数据权限 3:本部门数据权限
                         4:本部门及以下数据权限)
                    format: int32
                remark:
                    type: string
                    description: 备注
                menuIds:
                    type: array
                    items:
                        type: string
                    description: 菜单ID列表
...
        api.system.v1.DeleteRoleReply:
            type: object
            properties: {}
        api.system.v1.DeleteRoleRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 角色ID
...
        api.system.v1.GetRoleReply:
            type: object
            properties:
                role:
                    $ref: '#/components/schemas/api.system.v1.RoleInfo'
...
        api.system.v1.ListRoleReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.RoleInfo'
                    description: 角色列表
                total:
                    type: integer
                    description: 总数
                    format: int32
...
        api.system.v1.UpdateRoleMenusRequest:
            type: object
            properties:
                roleId:
                    type: string
                menuIds:
                    type: array
                    items:
                        type: string
...